backend-chatbot/requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.33.0
sqlalchemy==2.0.43
pymysql==1.1.2
python-dotenv==1.0.1
google-generativeai==0.1.0rc1
pydantic==2.10.6
starlette==0.44.0